# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /charities:
    get:
      operationId: listCharities
      tags: [charities]
      summary: List all charities
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Charities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Charity"
    post:
      operationId: createCharity
      tags: [charities]
      summary: Create a charity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityBody"
      responses:
        "201":
          description: Charity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
  /charities/{charityId}:
    get:
      operationId: getCharity
      tags: [charities]
      summary: Get a charity by ID
      parameters:
        - name: charityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charity found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
        "404":
          description: Charity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCharity
      tags: [charities]
      summary: Update a charity (admin)
      parameters:
        - name: charityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharityBody"
      responses:
        "200":
          description: Charity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
    delete:
      operationId: deleteCharity
      tags: [charities]
      summary: Delete a charity (admin)
      parameters:
        - name: charityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charity deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /charities/select:
    post:
      operationId: selectCharity
      tags: [charities]
      summary: Select a charity and set contribution percentage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectCharityBody"
      responses:
        "200":
          description: Charity selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCharitySelection"

  # Draws
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          minimum: 10
          maximum: 100
          default: 10
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          nullable: true
        subscriptionStatus:
          type: string
          enum: [active, inactive, cancelled, expired]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Charity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        featured:
          type: boolean
        totalContributions:
          type: number
        subscriberCount:
          type: integer
        upcomingEvents:
          type: array
          items:
            $ref: "#/components/schemas/CharityEvent"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - featured
        - totalContributions
        - subscriberCount
        - createdAt
    CharityEvent:
      type: object
      properties:
        id:
          type: integer
        charityId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date-time
      required:
        - id
        - charityId
        - title
        - eventDate
    CreateCharityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        featured:
          type: boolean
          default: false
      required:
        - name
        - description
    UpdateCharityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        featured:
          type: boolean
    UserCharitySelection:
      type: object
      properties:
        charity:
          $ref: "#/components/schemas/Charity"
          nullable: true
        charityId:
          type: integer
          nullable: true
        charityPercentage:
          type: number
          nullable: true
      required:
        - charityPercentage
    SelectCharityBody:
      type: object
      properties:
        charityId:
          type: integer
        charityPercentage:
          type: number
          minimum: 10
          maximum: 100
      required:
        - charityId
        - charityPercentage

    # Draw
...