# lib/api-zod/src/generated/types/charityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CharityEvent {
  id: number;
  charityId: number;
  title: string;
  description?: string | null;
  eventDate: Date;
}