# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserSubscriptionStatus } from "./userSubscriptionStatus";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  charityId?: number | null;
  charityPercentage?: number | null;
  subscriptionStatus?: UserSubscriptionStatus;
  createdAt: Date;
}